Add versioned stream native-handle shim; test via libtorch_agn_2_13
