+++ torch/csrc/stable/accelerator.h
@@ -92,12 +92,14 @@ class Stream {
     return stream_id;
   }
 
+#if TORCH_FEATURE_VERSION >= TORCH_VERSION_2_13_0
   void* nativeHandle() const {
     void* native_handle = nullptr;
     TORCH_ERROR_CODE_CHECK(
         aoti_torch_stream_native_handle(stream_.get(), &native_handle));
     return native_handle;
   }
+#endif // TORCH_FEATURE_VERSION >= TORCH_VERSION_2_13_0
 
  private:
   std::shared_ptr<StreamOpaque> stream_;
